finish the clean code of the f1 and strt the f2

+++ src/Game/game/model/Move/impactAble.java
@@ -15,7 +15,7 @@ public interface impactAble extends Moveable {
 
     void setSpeed (double speed);
 
-    double getSpeed ();
+    float getSpeed ();
 
     void setMoveDirection (Direction direction);
